Quantum-SVDD-PyTorch-master/requirements.txt
appdirs==1.4.4
argon2-cffi==21.1.0
attrs==21.2.0
autograd==1.3
autoray==0.2.5
backcall==0.2.0
bleach==4.1.0
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
Click==7.0
cryptography==37.0.2
cvxopt==1.2.7
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
dlx==1.0.4
docplex==2.22.213
entrypoints==0.3
fastdtw==0.3.4
fastjsonschema==2.15.1
fonttools==4.28.2
future==0.18.2
h5py==3.2.1
idna==3.3
immutables==0.6
importlib-resources==5.4.0
inflection==0.5.1
ipykernel==6.5.1
ipython==7.30.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.8.0
kiwisolver==1.3.2
lark==1.1.2
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.12.0
mpmath==1.2.1
msgpack==0.6.2
multitasking==0.0.10
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.7.1
ninja==1.10.2.3
notebook==6.4.6
ntlm-auth==1.5.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
PennyLane==0.22.1
PennyLane-Forest==0.24.0
PennyLane-Lightning==0.22.1
pickleshare==0.7.5
Pillow==7.0.0
ply==3.11
prometheus-client==0.12.0
prompt-toolkit==3.0.23
psutil==5.8.0
pycparser==2.21
pydot==1.4.2
Pygments==2.10.0
pylatexenc==2.10
pyparsing==3.0.6
pyquil==2.28.2
pyrsistent==0.18.0
python-constraint==1.4.0
python-dateutil==2.8.2
python-rapidjson==1.8
pytz==2021.3
pywin32==302
pywinpty==1.1.6
pyzmq==22.3.0
qiskit==0.32.1
qiskit-aer==0.9.1
qiskit-aqua==0.9.5
qiskit-ibmq-provider==0.18.1
qiskit-ignis==0.6.0
qiskit-terra==0.18.3
Quandl==3.7.0
requests==2.26.0
requests-ntlm==1.1.0
retworkx==0.10.2
rpcq==3.10.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scikit-opt==0.6.6
scipy==1.8.0
seaborn==0.11.2
semantic-version==2.6.0
Send2Trash==1.8.0
setuptools-scm==6.3.2
sympy==1.9
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
torch==1.11.0
torchaudio==0.11.0+cu113
torchvision==0.12.0
tornado==6.1
traitlets==5.1.1
tweedledum==1.1.1
ujson==1.35
urllib3==1.24.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.2
wincertstore==0.2
yfinance==0.1.67
zipp==3.6.0
